+++ src/pages/jaewoong/Main/Main.scss
@@ -120,9 +120,17 @@ main {
   display: flex;
   margin: 50px 250px;
 
-  .main-left {
-    width: 60%;
+  .main-box{
+    width: 100%;
     margin-right: 20px;
+    display: flex;
+    flex-direction: column;
+  }
+
+  .main-left {
+    width: 100%;
+    margin-right: 20px;
+    margin-bottom: 20px;
     border: 1px solid lightgray;
   }
 
@@ -151,7 +159,7 @@ main {
 
   .feed-img img {
     width: 100%;
-    height: 560px;
+    height: 600px;
   }
 
   .feed-content {
